+++ CHANGELOG.md
@@ -24,6 +24,7 @@ we hit release version 1.0.0.
 - `BrillouinZone.merge` allows simple merging of several objects, #537
 
 ### Changed
+- Added printout of the removed couplings in the `RecursiveSI`
 - SuperCell class is officially deprecated in favor of Lattice, see #95 for details
   The old class will still be accessible and usable for some time (at least a year)
 - Enabled EigenState.wavefunction(grid) to accept grid as the initialization of